interaction: introduce max_replay to achieve stateful services
This commit allows to build services that reply different responses
to the same request. This is achieved by introducing a `max_replay`
value which tells to the service how many times that interaction
can actually be played again. After this number is reached, the next
matching interaction is used to reply to the same request.

Signed-off-by: Antonio Cesarano <cesarano2607@gmail.com>